changed isOlderThan(int) to double

+++ app/src/main/java/info/nightscout/androidaps/db/CareportalEvent.java
@@ -98,7 +98,7 @@ public class CareportalEvent implements DataPointWithLabelInterface {
             return diff.get(TimeUnit.DAYS) + " " + MainApp.sResources.getString(R.string.days) + " " + diff.get(TimeUnit.HOURS) + " " + MainApp.sResources.getString(R.string.hours);
     }
 
-    public boolean isOlderThan(int threshold) {
+    public boolean isOlderThan(double threshold) {
         Map<TimeUnit, Long> diff = computeDiff(date, System.currentTimeMillis());
             if(diff.get(TimeUnit.DAYS) > threshold)
                 return true;
